Initial mobile style

+++ composables/text.ts
@@ -24,6 +24,6 @@ export const useShuffle = (
 export const useFormattedDate = (
   date: number,
   ) => {
-  const options = { year: 'numeric', month: 'long', day: 'numeric' } as any;
+  const options = { year: 'numeric', month: 'short', day: 'numeric' } as any;
   return new Date(date * 1000).toLocaleDateString("en-US", options);
 }
